# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WanisAI API - consent-first cognitive health companion
servers:
  - url: /api
    description: Base API path
...
paths:
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
        '500':
          description: Failed to generate upload URL.
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content.
        '404':
          description: Object not found.
...
  /medications:
    get:
      operationId: listMedications
      tags: [medications]
      summary: List all medications
      responses:
        "200":
          description: List of medications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: addMedication
      tags: [medications]
      summary: Add a medication manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationInput"
      responses:
        "201":
          description: Created medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /medications/{id}:
    delete:
      operationId: deleteMedication
      tags: [medications]
      summary: Remove a medication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
        "404":
          description: Not found
  /medications/extract-vision:
    post:
      operationId: extractMedicationVision
      tags: [medications]
      summary: Extract medication details from an uploaded image
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicationVisionExtractRequest"
      responses:
        "200":
          description: Extracted medication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicationVisionExtractResponse"
...
  /guardian-profile:
    get:
      operationId: getGuardianProfile
      tags: [guardian]
      summary: Get guardian/pilgrim safety profile
      responses:
        "200":
          description: Guardian profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuardianProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: upsertGuardianProfile
      tags: [guardian]
      summary: Create or update guardian profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuardianProfileInput"
      responses:
        "200":
          description: Updated guardian profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuardianProfile"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    GuardianProfile:
      type: object
      properties:
        id:
          type: integer
        pilgrimName:
          type: string
        pilgrimPhotoUrl:
          type: ["string", "null"]
        hotelName:
          type: string
        hotelAddress:
          type: string
        hotelPhone:
          type: ["string", "null"]
        groupLeaderName:
          type: string
        groupLeaderPhone:
          type: string
        medicalNotes:
          type: ["string", "null"]
        emergencyNote:
          type: ["string", "null"]
        meetingPointName:
          type: ["string", "null"]
        meetingPointAddress:
          type: ["string", "null"]
        emergencyContactPhone:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - pilgrimName
        - hotelName
        - hotelAddress
        - groupLeaderName
        - groupLeaderPhone
        - createdAt
        - updatedAt
    GuardianProfileInput:
      type: object
      properties:
        pilgrimName:
          type: string
          minLength: 1
        pilgrimPhotoUrl:
          type: string
        hotelName:
          type: string
          minLength: 1
        hotelAddress:
          type: string
          minLength: 1
        hotelPhone:
          type: string
        groupLeaderName:
          type: string
          minLength: 1
        groupLeaderPhone:
          type: string
          minLength: 1
        medicalNotes:
          type: string
        emergencyNote:
          type: string
        meetingPointName:
          type: string
        meetingPointAddress:
          type: string
        emergencyContactPhone:
          type: string
      required:
        - pilgrimName
        - hotelName
        - hotelAddress
        - groupLeaderName
        - groupLeaderPhone
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...
    Medication:
      type: object
      required: [id, name, dosage, frequency, acbScore, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        acbScore:
          type: integer
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    MedicationInput:
      type: object
      required: [name, dosage, frequency]
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        reason:
          type: string
          nullable: true
    MedicationVisionExtractRequest:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
    MedicationVisionExtractResponse:
      type: object
      required: [name, dosage, frequency, acbScore]
      properties:
        name:
          type: string
        dosage:
          type: string
        frequency:
          type: string
        acbScore:
          type: integer
...