# lib/api-zod/src/generated/types/medicationInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * WanisAI API - consent-first cognitive health companion
 * OpenAPI spec version: 0.1.0
 */

export interface MedicationInput {
  name: string;
  dosage: string;
  frequency: string;
  reason?: string | null;
}